feat: refactor API usage logging to include principal information and update response time type

+++ service/core/src/main/java/cloud/xcan/angus/core/ai/interfaces/application/facade/vo/ApplicationDetailVo.java
@@ -220,7 +220,7 @@ public class ApplicationDetailVo extends TenantAuditingVo {
     private Long totalTokens;
 
     @Schema(description = "平均响应时间（秒）")
-    private Double avgResponseTime;
+    private Long avgResponseTime;
 
     @Schema(description = "成功率")
     private Double successRate;
